public_api/requirements.txt
python-multipart
fastapi
uvicorn[standard]
sqlalchemy>=1.4
pika
requests
pydantic
httpx
PyMySQL
python-dotenv